Consolidate shared model code under model/ submodule
