+++ src/every_query/model/model.py
