+++ src/every_query/train/configs/_demo_train.yaml
@@ -21,7 +21,7 @@ query:
 
 datamodule:
   _target_: meds_torchdata.extensions.lightning_datamodule.Datamodule
-  data_class: every_query.dataset.EveryQueryPytorchDataset
+  data_class: every_query.model.dataset.EveryQueryPytorchDataset
   config:
     _target_: meds_torchdata.MEDSTorchDataConfig
     tensorized_cohort_dir: ???
@@ -35,7 +35,7 @@ datamodule:
   pin_memory: false
 
 lightning_module:
-  _target_: every_query.lightning_module.EveryQueryLightningModule
+  _target_: every_query.model.lightning_module.EveryQueryLightningModule
   model:
     _target_: every_query.model.EveryQueryModel
     precision: "32-true"
