+++ src/every_query/train/configs/config.yaml
@@ -12,7 +12,7 @@ query:
 
 datamodule:
   _target_: meds_torchdata.extensions.lightning_datamodule.Datamodule
-  data_class: dataset.EveryQueryPytorchDataset
+  data_class: every_query.model.dataset.EveryQueryPytorchDataset
   config:
     _target_: meds_torchdata.MEDSTorchDataConfig
     tensorized_cohort_dir: ${oc.env:FINAL_DATA_DIR}
@@ -25,7 +25,7 @@ datamodule:
   pin_memory: True
 
 lightning_module:
-  _target_: every_query.lightning_module.EveryQueryLightningModule
+  _target_: every_query.model.lightning_module.EveryQueryLightningModule
   model:
     _target_: every_query.model.EveryQueryModel
     precision: "16-mixed"
